+++ app/src/main/java/woowacourse/shopping/OnProductItemClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.cart
+package woowacourse.shopping
 
 fun interface OnProductItemClickListener {
     fun onClick(id: Int)
