+++ app/src/main/res/layout/holder_cart.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="onCartItemListener"
-            type="woowacourse.shopping.cart.OnProductItemClickListener" />
+            type="woowacourse.shopping.OnProductItemClickListener" />
 
     </data>
