+++ app/src/main/res/layout/holder_product.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="onProductItemClickListener"
-            type="woowacourse.shopping.cart.OnProductItemClickListener" />
+            type="woowacourse.shopping.OnProductItemClickListener" />
 
     </data>
